+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Change the upgrade to only upgrade Mac drivers if Apple hardware is detected.
 - Change the first-time setup to install drivers from the [extra] repository instead of [community] now that they have been merged.
 - Add Pacman support in the Discover package manager.
+- Change Snapper to only keep 10 hourly snapshots.
 
 ## 3.2.1 - The SteamOS 3.4 Update